User service - create User method

+++ src/main/java/em/fire/backend/service/user/impl/UserServiceImpl.java
@@ -11,7 +11,7 @@ import em.fire.backend.repository.UserJpaRepository;
 import em.fire.backend.service.user.UserService;
 
 @Service("userService")
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService {
 
 	@Autowired
 	@Qualifier("userJpaRepository")
@@ -19,11 +19,13 @@ public class UserServiceImpl implements UserService{
 
 	@Autowired
 	@Qualifier("userDSLRepository")
-	private UserDSLRepository userDSLRepository;	
-	
+	private UserDSLRepository userDSLRepository;
+
 	@Override
-	public User save(User user) {
+	public User createUser(User user) {
+
 		return userJpaRepository.save(user);
+
 	}
 
 	@Override
@@ -32,14 +34,19 @@ public class UserServiceImpl implements UserService{
 	}
 
 	@Override
-	public boolean existsById(Long id) {	
+	public boolean existsById(Long id) {
 		return userJpaRepository.existsById(id);
 	}
 
 	@Override
-	public void deleteUserById(Long id) {
-		userJpaRepository.deleteById(id);
+	public boolean deleteUserById(Long id) {
+
+		if (userJpaRepository.existsById(id)) {
+			userJpaRepository.deleteById(id);
+
+			return true;
+		}
+		return false;
 	}
-	
-	
+
 }
